+++ packages/cubejs-api-gateway/index.js
@@ -130,7 +130,8 @@ const querySchema = Joi.object().keys({
   timezone: Joi.string(),
   limit: Joi.number().integer().min(1).max(50000),
   offset: Joi.number().integer().min(0),
-  renewQuery: Joi.boolean()
+  renewQuery: Joi.boolean(),
+  ungrouped: Joi.boolean()
 });
 
 const normalizeQuery = (query) => {